Moved Attributes namespace to root
