+++ lib/Attributes/BeforeClassMethods.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace PhpBench\Benchmark\Metadata\Attributes;
+namespace PhpBench\Attributes;
+
+use Attribute;
 
 #[Attribute(Attribute::TARGET_CLASS | Attribute::TARGET_METHOD)]
 class BeforeClassMethods extends AbstractMethodsAttribute
