+++ lib/Attributes/OutputTimeUnit.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace PhpBench\Benchmark\Metadata\Attributes;
+namespace PhpBench\Attributes;
 
 use Attribute;
