+++ lib/Attributes/Skip.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace PhpBench\Attributes;
+
+use Attribute;
+
+#[Attribute(Attribute::TARGET_CLASS | Attribute::TARGET_METHOD)]
+final class Skip
+{
+}
