--- lib/Benchmark/Metadata/Attributes/AfterMethods.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace PhpBench\Benchmark\Metadata\Attributes;
-
-use Attribute;
-
-#[Attribute(Attribute::TARGET_CLASS | Attribute::TARGET_METHOD)]
-class AfterMethods extends AbstractMethodsAttribute
-{
-}
