+++ tests/Unit/Benchmark/Metadata/Driver/AttributeDriverTest.php
@@ -3,22 +3,22 @@
 namespace PhpBench\Tests\Unit\Benchmark\Metadata\Driver;
 
 use Generator;
-use PhpBench\Benchmark\Metadata\Attributes\AfterClassMethods;
-use PhpBench\Benchmark\Metadata\Attributes\AfterMethods;
-use PhpBench\Benchmark\Metadata\Attributes\Assert;
-use PhpBench\Benchmark\Metadata\Attributes\BeforeClassMethods;
-use PhpBench\Benchmark\Metadata\Attributes\BeforeMethods;
-use PhpBench\Benchmark\Metadata\Attributes\Executor;
-use PhpBench\Benchmark\Metadata\Attributes\Groups;
-use PhpBench\Benchmark\Metadata\Attributes\Iterations;
-use PhpBench\Benchmark\Metadata\Attributes\OutputMode;
-use PhpBench\Benchmark\Metadata\Attributes\OutputTimeUnit;
-use PhpBench\Benchmark\Metadata\Attributes\ParamProviders;
-use PhpBench\Benchmark\Metadata\Attributes\Revs;
-use PhpBench\Benchmark\Metadata\Attributes\Skip;
-use PhpBench\Benchmark\Metadata\Attributes\Sleep;
-use PhpBench\Benchmark\Metadata\Attributes\Timeout;
-use PhpBench\Benchmark\Metadata\Attributes\Warmup;
+use PhpBench\Attributes\AfterClassMethods;
+use PhpBench\Attributes\AfterMethods;
+use PhpBench\Attributes\Assert;
+use PhpBench\Attributes\BeforeClassMethods;
+use PhpBench\Attributes\BeforeMethods;
+use PhpBench\Attributes\Executor;
+use PhpBench\Attributes\Groups;
+use PhpBench\Attributes\Iterations;
+use PhpBench\Attributes\OutputMode;
+use PhpBench\Attributes\OutputTimeUnit;
+use PhpBench\Attributes\ParamProviders;
+use PhpBench\Attributes\Revs;
+use PhpBench\Attributes\Skip;
+use PhpBench\Attributes\Sleep;
+use PhpBench\Attributes\Timeout;
+use PhpBench\Attributes\Warmup;
 use PhpBench\Benchmark\Metadata\BenchmarkMetadata;
 use PhpBench\Benchmark\Metadata\Driver\AttributeDriver;
 use PhpBench\Benchmark\Metadata\SubjectMetadata;
